Fix dialog converter after tuvok API change.

+++ ImageVis3D/IO/DialogConverter.h
@@ -51,7 +51,7 @@ public:
 
   virtual bool ConvertToRAW(const std::string& strSourceFilename, 
                             const std::string& strTempDir, bool bNoUserInteraction,
-                            uint64_t& iHeaderSkip, uint64_t& iComponentSize, uint64_t& iComponentCount, 
+                            uint64_t& iHeaderSkip, unsigned& iComponentSize, uint64_t& iComponentCount, 
                             bool& bConvertEndianess, bool& bSigned, bool& bIsFloat, UINT64VECTOR3& vVolumeSize,
                             FLOATVECTOR3& vVolumeAspect, std::string& strTitle,
                             std::string& strIntermediateFile,
